+++ core/framework/include/Structs.h
@@ -176,7 +176,7 @@ namespace MyosotisFW
 				glm::mat4 view;
 				glm::mat4 projection;
 				glm::vec4 color;
-				float renderID;
+				uint32_t renderID;
 			}data;
 		}standardUBO;			// UBO
 	}StaticMeshShaderObject;
@@ -194,7 +194,7 @@ namespace MyosotisFW
 				glm::mat4 view;
 				glm::mat4 projection;
 				glm::vec4 color;
-				float renderID;
+				uint32_t renderID;
 			}data;
 		}standardUBO;			// UBO
 	}SkyboxShaderObject;
@@ -215,7 +215,7 @@ namespace MyosotisFW
 				glm::mat4 view;
 				glm::mat4 projection;
 				glm::vec4 color;
-				float renderID;
+				uint32_t renderID;
 			}data;
 		}standardUBO;			// UBO
 	}InteriorObjectShaderObject;
